refactor cpu card playing
